backend/requirements.txt
pytest==6.0.1
Flask==1.1.2
pubnub==4.5.3
python-decouple==3.3
